Move the StudentLoansClaimsCsv into the StudentLoans module
