+++ app/controllers/admin/claims_controller.rb
@@ -5,7 +5,7 @@ class Admin::ClaimsController < Admin::BaseAdminController
     @claims = Claim.includes(eligibility: [:claim_school, :current_school]).awaiting_approval.order(:submitted_at)
 
     respond_to do |format|
-      format.csv { send_file StudentLoansClaimsCsv.new(@claims).file, type: "text/csv", filename: "claims.csv" }
+      format.csv { send_file StudentLoans::ClaimsCsv.new(@claims).file, type: "text/csv", filename: "claims.csv" }
       format.html
     end
   end
