+++ spec/models/student_loans/claims_csv_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe StudentLoansClaimsCsv do
+RSpec.describe StudentLoans::ClaimsCsv do
   before do
     create(:claim, :submitted)
   end
